Create YMCA
Young's Modulus CSV analyser
Starrett FMS 500 produces numbered .csv files.
This program, when placed in the same folder as said files:
Scans columns for relevant data, exports into a new file, plots the data, and analyses for the fracture point and the linear fit of the slope just before. This is then used to provide the Young's modulus of compression in MPa. Finally, the new file is deleted, and the plot and data is saved.